removed pre events
