+++ src/com/palmergames/bukkit/towny/event/NationRemoveAllyEvent.java
@@ -1,19 +1,37 @@
 package com.palmergames.bukkit.towny.event;
 
 import com.palmergames.bukkit.towny.object.Nation;
+import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
 
-public class NationRemoveAllyEvent extends Event {
+public class NationRemoveAllyEvent extends Event implements Cancellable {
+
 	private static final HandlerList handlers = new HandlerList();
 	private final Nation nation;
 	private final Nation removedNation;
+	boolean cancelled;
+	private String cancelMessage = "Sorry this event was cancelled";
 
 	public NationRemoveAllyEvent(Nation nation, Nation removedNation) {
 		this.nation = nation;
 		this.removedNation = nation;
 
 	}
+
+	public boolean isCancelled() {
+		return cancelled;
+	}
+
+	public void setCancelled(boolean cancel) {
+		cancelled = cancel;
+	}
+
+	public void setCancelMessage(String cancelMessage) {
+		this.cancelMessage = cancelMessage;
+	}
+	
+	public String getCancelMessage() { return this.cancelMessage; }
 
 	public Nation getRemovedNation() {
 		return removedNation;
@@ -31,6 +49,6 @@ public class NationRemoveAllyEvent extends Event {
 	public HandlerList getHandlers() {
 		return handlers;
 	}
-	
-	
+
+
 }
